+++ control/src/input/mod.rs
@@ -4,4 +4,3 @@ pub mod control;
 pub mod pot;
 pub mod snapshot;
 pub mod store;
-pub mod switch;