+++ control/src/input/switch.rs
@@ -1,4 +1,0 @@
-//! Switch is a simple on/off slider.
-
-/// Switch on or off.
-pub type Switch = bool;